Changing normal alert into sweet alert with confirmation sweet alert for data deletion

+++ deletesurvey1.php
@@ -1,11 +1,18 @@
 <?php
 include "sambung.php";
 session_start();
-$_SESSION['status'] = "Data telah berjaya dipadam!";
-$_SESSION['status_code'] = "success";
+
 $id = $_GET['id_delete'];
 
-$padam = mysqli_query($hubung, "DELETE FROM entrance WHERE idEnt = '$id'"); 
-	echo "<script>
-	      window.location = 'total.php'</script>";
+if ($id != null) {
+
+	$padam = mysqli_query($hubung, "DELETE FROM entrance WHERE idEnt = '$id'"); 
+	$_SESSION['status'] = "Data telah berjaya dipadam!";
+	$_SESSION['status_code'] = "success";
+	echo "<script>window.location = 'total.php'</script>";
+
+} else {
+	$_SESSION['status'] = "Ralat telah terjadi. Sila rujuk admin sistem.";
+	$_SESSION['status_code'] = "error";
+}
 ?>
